requirements.txt
prefect 
duckdb
polars
pandas
streamlit
python-dotenv